# docs/tech/openapi.yaml
openapi: 3.1.0
info:
  title: Tassl API
  version: 1.0.0
  description: REST API for Tassl, the AI decision simulator for higher education. Generated from the Zod schemas in src/server/modules/*/schema.ts by `pnpm openapi:generate`. Every response carries an `x-request-id` header. Errors use the envelope in `#/components/schemas/ErrorEnvelope`.
servers:
  - url: /api/v1
    description: Same-origin, version 1
...
security:
  - cookieAuth: []
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    cookieAuth:
      type: apiKey
      in: cookie
      name: better-auth.session_token
    cronAuth:
      type: http
      scheme: bearer
      description: CRON_SECRET
  parameters:
...
    cursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
...
    limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    packageId:
      name: packageId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    runId:
      name: runId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    Forbidden:
      description: Not permitted
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    NotFound:
      description: Missing or cross-tenant
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    RateLimited:
      description: Rate limited
      headers:
        Retry-After:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Unauthenticated:
      description: No session
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
...
    Clock:
      type: object
      required:
        - remainingMs
        - paused
        - creditedMs
        - chargedMs
      properties:
        remainingMs:
          type: integer
        paused:
          type: boolean
        creditedMs:
          type: integer
        chargedMs:
          type: integer
...
    ErrorEnvelope:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
            - requestId
          properties:
            code:
              type: string
              pattern: ^[A-Z0-9_]+$
            message:
              type: string
            details: {}
            requestId:
              type: string
...
    Notification:
      type: object
      required:
        - id
        - type
        - title
        - body
        - link
        - readAt
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum:
            - generation_complete
            - generation_failed
            - run_scored
            - run_held
            - bands_confirmed
            - invitation
            - export_ready
        title:
          type: string
        body:
          type: string
        link:
          type:
            - string
            - "null"
        readAt:
          type:
            - string
            - "null"
          format: date-time
        createdAt:
          type: string
          format: date-time
...
    PackageView:
      type: object
      required:
        - id
        - title
        - familyKey
        - versions
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        familyKey:
          type: string
        versions:
          type: array
          items:
            type: object
            required:
              - id
              - version
              - status
            properties:
              id:
                type: string
                format: uuid
              version:
                type: integer
              status:
                type: string
    Page:
      type: object
      required:
        - items
        - nextCursor
      properties:
        items:
          type: array
          items: {}
        nextCursor:
          type:
            - string
            - "null"
...
    RunState:
      type: string
      enum:
        - assigned
        - readiness
        - framing
        - working
        - paused
        - decision_locked
        - turn_open
        - turn_locked
        - defense_pending
        - defense_complete
        - scored
        - confirmed
        - recorded
        - voided
        - abandoned
        - defense_missed
        - under_appeal
        - expired
    RunSummary:
      type: object
      required:
        - id
        - assignmentId
        - attemptNo
        - state
        - mode
        - variantKey
        - isWalkthrough
        - clock
        - turn
        - scoringStatus
        - timestamps
        - links
      properties:
        id:
          type: string
          format: uuid
        assignmentId:
          type: string
          format: uuid
        attemptNo:
          type: integer
        state:
          $ref: "#/components/schemas/RunState"
        mode:
          type: string
          enum:
            - guided
            - standard
            - open
        variantKey:
          type: string
          enum:
            - defective
            - sound
        isWalkthrough:
          type: boolean
        clock:
          oneOf:
            - $ref: "#/components/schemas/Clock"
            - type: "null"
        turn:
          type:
            - object
            - "null"
          properties:
            dueAt:
              type: string
              format: date-time
            windowEndsAt:
              type:
                - string
                - "null"
              format: date-time
        scoringStatus:
          type: string
          enum:
            - idle
            - queued
            - running
            - held
            - done
        timestamps:
          type: object
          additionalProperties:
            type:
              - string
              - "null"
            format: date-time
        links:
          type: object
          required:
            - next
          properties:
            next:
              type: string
...
paths:
...
  /courses/{courseId}:
    get:
      tags:
        - courses
      operationId: getCourse
      summary: Course detail
      parameters:
        - in: path
          name: courseId
          schema:
            type: string
            format: uuid
            pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
          required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                    pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                  organizationId:
                    type: string
                  name:
                    type: string
                  term:
                    type: string
                  outsideAiPolicy:
                    type: string
                    enum:
                      - open
                      - declared
                      - in_environment_only
                  mapping:
                    type: object
                    properties:
                      novice:
                        type: number
                        exclusiveMinimum: 0
                      developing:
                        type: number
                        exclusiveMinimum: 0
                      proficient:
                        type: number
                        exclusiveMinimum: 0
                      professional:
                        type: number
                        exclusiveMinimum: 0
                    required:
                      - novice
                      - developing
                      - proficient
                      - professional
                    additionalProperties: false
                  defaultRunWeight:
                    type: number
                  critiqueWeightFactor:
                    type: number
                  taughtConcepts:
                    type: array
                    items:
                      type: string
                  sections:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          format: uuid
                          pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                        courseId:
                          type: string
                          format: uuid
                          pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                        name:
                          type: string
                        memberCount:
                          type: integer
                          minimum: 0
                          maximum: 9007199254740991
                        assignmentCount:
                          type: integer
                          minimum: 0
                          maximum: 9007199254740991
                      required:
                        - id
                        - courseId
                        - name
                        - memberCount
                        - assignmentCount
                      additionalProperties: false
                  assignments:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          format: uuid
                          pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                        sectionId:
                          type: string
                          format: uuid
                          pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                        label:
                          type: string
                        runType:
                          type: string
                          enum:
                            - decision
                            - critique
                        packageVersionId:
                          type: string
                          format: uuid
                          pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                        variantId:
                          type: string
                          format: uuid
                          pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                        workingClockSeconds:
                          anyOf:
                            - type: integer
                              minimum: -9007199254740991
                              maximum: 9007199254740991
                            - type: "null"
                        weight:
                          type:
                            - number
                            - "null"
                        isWalkthrough:
                          type: boolean
                        opensAt:
                          anyOf:
                            - type: string
                              format: date-time
                              pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d:[0-5]\d(?:\.\d+)?(?:Z))$
                            - type: "null"
                      required:
                        - id
                        - sectionId
                        - label
                        - runType
                        - packageVersionId
                        - variantId
                        - workingClockSeconds
                        - weight
                        - isWalkthrough
                        - opensAt
                      additionalProperties: false
                required:
                  - id
                  - organizationId
                  - name
                  - term
                  - outsideAiPolicy
                  - mapping
                  - defaultRunWeight
                  - critiqueWeightFactor
                  - taughtConcepts
                  - sections
                  - assignments
                additionalProperties: false
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
    patch:
      tags:
        - courses
      operationId: updateCoursePolicy
      summary: Update policy, weights, taught concepts
      parameters:
        - in: path
          name: courseId
          schema:
            type: string
            format: uuid
            pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 200
                term:
                  type: string
                  minLength: 1
                  maxLength: 100
                outsideAiPolicy:
                  type: string
                  enum:
                    - open
                    - declared
                    - in_environment_only
                mapping:
                  type: object
                  properties:
                    novice:
                      type: number
                    developing:
                      type: number
                    proficient:
                      type: number
                    professional:
                      type: number
                  required:
                    - novice
                    - developing
                    - proficient
                    - professional
                defaultRunWeight:
                  type: number
                  minimum: 0
                critiqueWeightFactor:
                  type: number
                  minimum: 0
                taughtConcepts:
                  maxItems: 50
                  type: array
                  items:
                    type: string
                    minLength: 1
                    maxLength: 120
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                    pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                  organizationId:
                    type: string
                  name:
                    type: string
                  term:
                    type: string
                  outsideAiPolicy:
                    type: string
                    enum:
                      - open
                      - declared
                      - in_environment_only
                  mapping:
                    type: object
                    properties:
                      novice:
                        type: number
                        exclusiveMinimum: 0
                      developing:
                        type: number
                        exclusiveMinimum: 0
                      proficient:
                        type: number
                        exclusiveMinimum: 0
                      professional:
                        type: number
                        exclusiveMinimum: 0
                    required:
                      - novice
                      - developing
                      - proficient
                      - professional
                    additionalProperties: false
                  defaultRunWeight:
                    type: number
                  critiqueWeightFactor:
                    type: number
                  taughtConcepts:
                    type: array
                    items:
                      type: string
                required:
                  - id
                  - organizationId
                  - name
                  - term
                  - outsideAiPolicy
                  - mapping
                  - defaultRunWeight
                  - critiqueWeightFactor
                  - taughtConcepts
                additionalProperties: false
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
...
  /notifications:
    get:
      tags:
        - notifications
      operationId: listNotifications
      summary: My notifications
      parameters:
        - $ref: "#/components/parameters/cursor"
        - $ref: "#/components/parameters/limit"
        - name: unread
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Page"
                  - type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: "#/components/schemas/Notification"
...
  /packages/{packageId}:
    get:
      tags:
        - packages
      operationId: getPackage
      summary: Package with versions
      parameters:
        - $ref: "#/components/parameters/packageId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PackageView"
...
  /runs/{runId}:
    get:
      tags:
        - runs
      operationId: getRun
      summary: Run status, clock, and timers
      parameters:
        - $ref: "#/components/parameters/runId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunSummary"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags:
        - courses
      operationId: deleteWalkthroughRun
      summary: Delete a walkthrough run
      parameters:
        - in: path
          name: runId
          schema:
            type: string
            format: uuid
            pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
          required: true
      responses:
        "204":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties: {}
                additionalProperties: false
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
...